Improve search for dependent packages in hip-lang
Use the installation prefix of the `hip-lang` package and the environment
value `ROCM_ROOT` (for consistency) as a hint to find `AMDDeviceLib` and
`amd_comgr`.
The path to `hip-lang-config.cmake` might not be part of the package
search paths in cmake, because cmake additionally checks `hipconfig`s
and `clang`s output to find it.
When this is the case help cmake to find these packages assuming they
share the same installation prefix as `hip-lang` (commonly `/opt/rocm`).